Implement parallel fuzzing
It is possible to run multiple fuzzing processes at the same time, thus
improving fuzzing speed signinficantly. One problem with the built-in
fuzzing is that it is bad at displaying information on the go.

Therefore we implement our own process mangement (using the awesome
tokio) and print results from all of them on the fly.