Remove some LLVM options which excessively slow down code
This commit is a bit of a stab towards #216. I discussed this a bit more
on google/oss-fuzz#3944 and this is intended to sync Rust's `cargo fuzz`
with the default C++ fuzzing there. Apparently the `prune-blocks` and
`trace-geps` options aren't used by libFuzzer and are either
experimental or disable too many optimizations.