Allow stripping of dead code
The `-Clink-dead-code` flag was added to fix an error with certain
targets.

However, it can also cause problems.

Example: You depend on crate A which links to library B. A includes an
unused reference to `B::foo`. If the specific version of B we have is
missing `foo`, we now get a linker error whereas we would not if
stripping dead code.