perf(coverage): Improve coverage collection by using merge
This approach creates a dummy-corpus and merges the main
corpus into the dummy-corpus. After the corpus is merged
we delete the dummy-corpus. This allows us to calculate
code coverage signficantly faster than calculating it
per-file in the corpus, as it can be run in parralell and
there is less startup initialisation cost.