Do not list dependency packages by MetadataCommand
When following `cargo-fuzz` tutorial at
https://rust-fuzz.github.io/book/cargo-fuzz/tutorial.html,
after executing `cargo fuzz init`, `fuzz/Cargo.toml`
contains entry `[dependencies.matches]` instead of
`[dependencies.url]`. This is caused by `MetadataCommand`
listing dependency packages in addition to main packages
which is not desired.

This does work as expected if result of `MetadataCommand`
contains the first package as the non-dep package but that does
not seem to work all the time.