Add `--codegen-units` flag.
The default in non-dev builds is 1, which gives better fuzzing
performance at a very high build-time cost, because it eliminates all
parallelism from the LLVM backend. There are times, e.g. short "smoke
test" fuzzing runs, where a faster build with slower fuzzing is a good
trade-off, and this new flag makes that easier to do.