Revert part of #415.
The second commit in #415 replaced the use of
`--config profile.release.debug="line-tables-only"` with
`-Cdebuginfo=line-tables-only`. I mistakenly thought these were
equivalent, but the former can also affect Cargo's use/non-use of
`-Csplit-debuginfo=unpacked` and `-Cstrip=debuginfo`.

As a result, cargo is now being run with the nonsensical combination of
`-Cdebuginfo=line-tables-only` and `-Cstrip=debuginfo`. I.e. "please
give me some debuginfo" and "please strip all debuginfo".

This commit reverts the relevant part of that commit and adds a test
that would have caught the problem. Sorry for the mix-up!